Change packages from eo.* to org.polystat.odin.*
