--- core/src/main/scala/eo/analysis/package.scala
@@ -1,24 +1,0 @@
-package eo
-
-//import eo.core.ast.EOProg
-//import eo.analysis.errors._
-
-//import cats.data.Validated
-
-/** Contains functions that analyze an EO program.
-  */
-package object analysis {
-//  decoration on top level
-
-//  attributes of defined objects are resolved
-//  def noUnresolvedNames(prog: EOProg): Validated[List[UnresolvedNameError], EOProg] = ???
-
-//  multiple decorations
-//  def noBindingRedefinition(prog: EOProg): Validated[List[RedefinitionError], EOProg] = ???
-
-//  name shadowing
-
-//  validBindings (e.g. no "> @" bindings at top level)
-
-//  no duplicated metas
-}
