--- core/src/main/scala/eo/core/ast/Constants.scala
@@ -1,19 +1,0 @@
-package eo.core.ast
-
-object Constants {
-  object ATTRS {
-    val DECORATION: String = "@"
-    val PARENT: String = "^"
-    val THIS: String = "$"
-  }
-
-  object SYMBS {
-    val BINDING: String = ">"
-    val CONST_MOD: String = "!"
-    val VARARG_MOD: String = "..."
-    val ARRAY_START: String = "*"
-    val FREE_ATTR_DECL_ST: String = "["
-    val FREE_ATTR_DECL_ED: String = "]"
-    val META_PREFIX: String = "+"
-  }
-}
