--- core/src/main/scala/eo/core/ast/astparams.scala
@@ -1,7 +1,0 @@
-package eo.core.ast
-
-import higherkindness.droste.data.Fix
-
-object astparams {
-  type EOExprOnly = Fix[EOExpr]
-}
