+++ core/src/main/scala/org/polystat/odin/analysis/package.scala
@@ -1,0 +1,24 @@
+package org.polystat.odin
+
+//import EOProg
+//import eo.analysis.errors._
+
+//import cats.data.Validated
+
+/** Contains functions that analyze an EO program.
+  */
+package object analysis {
+//  decoration on top level
+
+//  attributes of defined objects are resolved
+//  def noUnresolvedNames(prog: EOProg): Validated[List[UnresolvedNameError], EOProg] = ???
+
+//  multiple decorations
+//  def noBindingRedefinition(prog: EOProg): Validated[List[RedefinitionError], EOProg] = ???
+
+//  name shadowing
+
+//  validBindings (e.g. no "> @" bindings at top level)
+
+//  no duplicated metas
+}
