+++ core/src/main/scala/org/polystat/odin/core/ast/astparams.scala
@@ -1,0 +1,7 @@
+package org.polystat.odin.core.ast
+
+import higherkindness.droste.data.Fix
+
+object astparams {
+  type EOExprOnly = Fix[EOExpr]
+}
