--- parser/src/main/scala/eo/parser/errors.scala
@@ -1,8 +1,0 @@
-package eo.parser
-
-
-object errors {
-  sealed trait ParsingError
-  sealed case class LexerError(msg: String) extends ParsingError
-  sealed case class ParserError(msg: String) extends ParsingError
-}
