+++ parser/src/main/scala/org/polystat/odin/parser/errors.scala
@@ -1,0 +1,7 @@
+package org.polystat.odin.parser
+
+object errors {
+  sealed trait ParsingError
+  sealed case class LexerError(msg: String) extends ParsingError
+  sealed case class ParserError(msg: String) extends ParsingError
+}
