+++ utils/src/main/scala/org/polystat/odin/utils/text.scala
@@ -1,0 +1,9 @@
+package org.polystat.odin.utils
+
+object text {
+  def indentWith(c: Char)(n: Int)(s: String) = s"${c.toString * n}${s}"
+
+  def indentSpace: Int => String => String = indentWith(' ')
+
+  def indent: String => String = indentWith(' ')(2)
+}
